\name{wastee-package}
\alias{wastee-package}
\alias{wastee}
\docType{package}
\title{
\packageTitle{wastee}
}
\description{
\packageDescription{wastee}
}
\details{
\tabular{ll}{
Package: \tab wastee\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2022-05-5\cr
License: \tab GPL (>= 2)\cr
}
}
\references{
    Andrews, D. W. K. and Ploberger, W. (1994). Optimal tests when a nuisance parameter is
    present only under the alternative. Econometrica, 62(6):1383-1414.

    Davies, R. B. (1977). Hypothesis testing when a nuisance parameter is present only under the
    alternative. Biometrika, 64(2):247-254.

    Davies, R. B. (1987). Hypothesis testing when a nuisance parameter is present only under the
    alternative. Biometrika, 74(1):33-43.

    Fan, A., Rui, S., and Lu, W. (2017). Change-plane analysis for subgroup detection and sample
    size calculation. Journal of the American Statistical Association, 112(518):769-778.

    Huang, Y., Cho, J., and Fong, Y. (2021). Threshold-based subgroup testing in logistic regression
    models in two phase sampling designs. Journal of the Royal Statistical Society: Series C. 291-311.

    LEE, S., SEO, M. H. and SHIN, Y. (2011). Testing for Threshold Effects in Regression Models. Journal of
    the American Statistical Association 106, 220-231.

    Liu, X. (2022). Testing under loss of identifiability. Manuscript.
}

\keyword{ package }
